Added controller support for the talent manager #4
Thanks to @wuergrob for the initial script/PR.

The Talent UI is a bit of a mess in controller/gamepad mode, so additional functionality is required to make it work properly with new talent entries. Additionally id/name is reversed in the controller UI, so a tweak was needed so the racial talent and backstab talent features work.